requirements.txt
openai
streamlit_chat
python-dotenv
